# employeeportal-client/src/app/app-service/model/employee.ts
/**
 * Employee Protal API
 * Manage the Employees of The Company
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export class Employee { 
    /**
     * ID of the Employee
     */
    employeeId?: number;
    /**
     * First Name of the Employee
     */
    firstName?: string;
    /**
     * Last Name of the Employee
     */
    lastName?: string;
    /**
     * Gender of the Employee
     */
    gender?: string;
    /**
     * Date of Birth of the Employee
     */
    dob?: string;
    /**
     * Department of the Employee
     */
    department?: string;
}