# employeeportal-service/src/main/java/io/swagger/api/EmployeeApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.5).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import javax.validation.Valid;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;
import io.swagger.model.Employee;
import io.swagger.model.Error;

@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2019-06-24T04:53:56.519Z")

@Api(value = "employee", description = "the employee API")
public interface EmployeeApi {

	@ApiOperation(value = "Registration of the Employee", nickname = "registerEmployee", notes = "Register an employee in the Employee Portal", response = String.class, tags = {})
	@ApiResponses(value = { @ApiResponse(code = 200, message = "Employee is registered", response = String.class),
			@ApiResponse(code = 409, message = "Employee already Exist", response = String.class),
			@ApiResponse(code = 400, message = "Fill mandatory fields", response = String.class),
			@ApiResponse(code = 200, message = "Unexpected error", response = Error.class) })
	@RequestMapping(value = "/employee", produces = { "application/json" }, method = RequestMethod.POST)
	ResponseEntity<String> registerEmployee(
			@ApiParam(value = "Employee Details", required = true) @Valid @RequestBody Employee employee);

}